+++ CHANGELOG.md
@@ -1,13 +1,48 @@
 # Braintree iOS SDK Release Notes
 
-## unreleased
+## 6.28.0 (2025-02-05)
+* BraintreeVenmo
+  * Allow universal links to be set without a return URL scheme (fixes #1505)
+* BraintreeCore
+  * Update to use `NSSecureCoding` protocol (fixes #1508)
+* BraintreePayPal
+  * Add `shopperSessionID` to `BTPayPalCheckoutRequest` and `BTPayPalVaultRequest`
+* BraintreeShopperInsights (BETA)
+  * Add `shopperSessionID` to `BTShopperInsightsClient` initializer
+  * Add `isPayPalAppInstalled()` and/or `isVenmoAppInstalled()`
+  * Replace `sendPayPalPresentedEvent()` and `sendPayPalPresentedEvent()` with `sendPresentedEvent(for:presentmentDetails:)`
+  * Add values to the following parameters to `presentmentDetails`:
+    * `experimentType`
+    * `pageType`
+    * `buttonOrder`
+  * Replace `sendPayPalSelectedEvent()` and `sendPayPalSelectedEvent()` with `sendSelectedEvent(for:)`
+
+## 6.27.0 (2025-01-23)
+* BraintreePayPal
+  * Add `BTContactInformation` request object
+  * Add `BTPayPalCheckoutRequest.contactInformation` optional property
+
+## 6.26.0 (2025-01-21)
+* BraintreePayPal
+  * Fix bug to ensure that `BTPayPalVaultRequest.userAuthenticationEmail` is not sent as an empty string
+  * Add `shippingCallbackURL` to `BTPayPalCheckoutRequest`
+* BraintreeThreeDSecure
+  * Return error if no `dfReferenceId` is returned in the 3D Secure flow
+
+## 6.25.0 (2024-12-11)
 * BraintreePayPal
   * Add `BTPayPalRequest.userPhoneNumber` optional property
   * Add PayPal App Switch checkout Flow (BETA)
     * Add `BTPayPalCheckoutRequest(userAuthenticationEmail:enablePayPalAppSwitch:amount:intent:userAction: offerPayLater:currencyCode:requestBillingAgreement:)`
     * **Note:** This feature is currently in beta and may change or be removed in future releases.
+  * Send `url` in `event_params` for App Switch events to PayPal's analytics service (FPTI)
 * BraintreeVenmo
   * Send `url` in `event_params` for App Switch events to PayPal's analytics service (FPTI)
+  * Add `BTVenmoClient(apiClient:universalLink:)` to use Universal Links when redirecting back from the Venmo flow
+* BraintreeCore
+  * Deprecate `BTAppContextSwitcher.sharedInstance.returnURLScheme`
+* BraintreeThreeDSecure
+  * Add `BTThreeDSecureRequest.requestorAppURL`
 
 ## 6.24.0 (2024-10-15)
 * BraintreePayPal